requirements.txt
fastapi[standard]
uvicorn
fastapi-health
azure-ai-projects
azure-ai-inference
azure-search-documents
pandas
python-dotenv
langchain-openai